% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkm-PCA.R
\name{findDistinguishingPCs}
\alias{findDistinguishingPCs}
\title{Find distinguishing PCs}
\usage{
findDistinguishingPCs(rotatedData, classLabels)
}
\arguments{
\item{rotatedData}{The samples' projections on PC space. Include here as many PCs as you want to inspect}

\item{classLabels}{Either a data frame (first column containing sample labels, second containing classes), or a list where the names of the list are the sample labels and the entries in the list are the classes.}
}
\value{
Returns a data.frame containing the following columns: PC (which PC was used for each comparison), AUROC (the area under the ROC curve for how well each PC distinguishes each class), P (rank sum p-value for how well the PC distinguishes the class), class (the class).
}
\description{
Identifies which PCs distinguish a feature of interest. If the feature of interest contains two classes, performs a single rank sum test for each PC.
If the feature of interest has >2 classes, performs a rank sum test to see if the PC separates the class from all others. 
In either case, returns a data frame containing the PC, class, AUROC, and rank sum P-value for how well the PC identifies the class.
}
\examples{
kmerMat = inputKMerFreqs(sprintf("kMerFiles/\%s.freq.gz",sampleDesc$id), IDs = sampleDesc$id)
myPCA = doKMerPCA(kmerMat, nPCs = "jackstraw")
}
