% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkm-PCA.R
\name{inputKMerFreqs}
\alias{inputKMerFreqs}
\title{Inputs k-mer frequencies}
\usage{
inputKMerFreqs(fileNames, IDs = fileNames)
}
\arguments{
\item{fileNames}{List of file names to input.}

\item{IDs}{List of sample IDs, same order as the file names. Defaults to fileNames.}
}
\value{
Returns a numeric matrix of kmer frequencies, kmers (rows) by samples (columns).
}
\description{
Inputs k-mer frequency data from a given set of file names. Each file is tab delimited and has the header: "kMer\\tPct_FG"
Followed by, for example "TTGATTG\\t0.012" ("<kMer>\\t<frequency>").
Assumes that all inputs have the same k-mer order.
}
\examples{
kmerMat = inputKMerFreqs(sprintf("kMerFiles/\%s.freq.gz",sampleDesc$id), IDs = sampleDesc$id)
}
