% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-mer_enrichment.R
\name{makeEnrichmentGraphForPC}
\alias{makeEnrichmentGraphForPC}
\title{Make a graph showing enrichment of a motif within a PC}
\usage{
makeEnrichmentGraphForPC(PC, binaryData, n_max = 3000, decreasing = F)
}
\arguments{
\item{PC}{A named numeric vector representing the k-mer loadings of a PC}

\item{binaryData}{A named logical vector representing k-mer cognate (T) or non-cognate (F) status for a motif.}

\item{n_max}{The maximum number of k-mers to consider for each minimum hypergeometric test. Defaults to 3000. Should be no more than about 10\% of the number of k-mers.}
}
\value{
Returns a list containing the plot (plot), the data.frame used to make the plot (rawData), and the minHG test (minHGTest).
}
\description{
Applies the minimum hypergeometric test to the current data, and makes a plot where the
}
\examples{
kmerMat = inputKMerFreqs(sprintf("kMerFiles/\%s.freq.gz",sampleDesc$id), IDs = sampleDesc$id)
myPCA = doKMerPCA(kmerMat, nPCs = "jackstraw")
treatmentPCs = findDistinguishingPCs(myPCA$rotation[,1:myPCA$nPCs], sampleDesc[c("id","treated")])
tfEnrichmentsPBM = getKMerTFEnrichment(myPCA$rotation[,1:myPCA$nPCs], cisbp$binaryPBMZScores);
tfEnrichmentsPBM = tfEnrichmentsPBM[order(tfEnrichmentsPBM$p),]
tfEnrichmentsPBM = merge(tfEnrichmentsPBM2, cisbp$TFTable[c("Motif_ID","TF_Name")], by="Motif_ID") # add TF names
p = makeEnrichmentGraphForPC(pcs$rotation[,treatmentPCs$PC[1]],cisbp$binaryPBMZScores[,head(tfEnrichmentsPBM$Motif_ID[tfEnrichmentsPBM$PC==treatmentPCs$PC[1] & tfEnrichmentsPBM$direction=="low"],n=1)]) #top motif for top treatment-distinguishing PC for lowly-weighted k-mers
}
